requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.3
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.9.0
pkg-resources==0.0.0
python-decouple==3.1
pytz==2018.7
static3==0.7.0
